+++ tests/tests_pytorch/deprecated_api/test_remove_2-0.py
@@ -315,16 +315,6 @@ def test_tuning_trainer_property():
         trainer.tuning = True
 
 
-@RunIf(skip_windows=True)
-def test_v1_8_0_deprecated_all_gather_grad():
-    tensor1 = torch.ones(1, requires_grad=True)
-    with mock.patch("torch.distributed.all_gather"), mock.patch("torch.distributed.get_world_size", return_value=1):
-        from pytorch_lightning.utilities import AllGatherGrad
-
-        with pytest.deprecated_call(match="`AllGatherGrad` has been deprecated in v1.8"):
-            AllGatherGrad.apply(tensor1)
-
-
 def test_v1_8_1_deprecated_rank_zero_only():
     from pytorch_lightning.utilities.distributed import rank_zero_only